Done the data validation on the login and signup added the logout funtionality added the data handling capabilities improved the status code and error from the backend server make webapp less crash prone

+++ client/src/service/api.js
@@ -37,6 +37,8 @@ axiosInstance.interceptors.response.use(
         return processResponse(response);
     },
     function(error) {
+        //if status code is not 2xx
+        console.error(error.response.body);
         // Stop global loader here
         return Promise.reject(ProcessError(error));
     }
@@ -53,7 +55,7 @@ const processResponse = (response) => {
         return {
             isFailure: true,
             status: response?.status,
-            msg: response?.msg,
+            msg: response?.data.msg,
             code: response?.code
         }
     }
@@ -65,29 +67,30 @@ const processResponse = (response) => {
 //////////////////////////////
 const ProcessError = async (error) => {
     if (error.response) {
+
         // Request made and server responded with a status code 
         // that falls out of the range of 2xx
         if (error.response?.status === 403) {
-            // const { url, config } = error.response;
-            // console.log(error);
-            // try {
-            //     let response = await API.getRefreshToken({ token: getRefreshToken() });
-            //     if (response.isSuccess) {
+            const { url, config } = error.response;
+            console.log(error);
+            try {
+                let response = await API.getRefreshToken({ token: getRefreshToken() });
+                if (response.isSuccess) {
                     sessionStorage.clear();
-            //         setAccessToken(response.data.accessToken);
+                    setAccessToken(response.data.accessToken);
 
-            //         const requestData = error.toJSON();
+                    const requestData = error.toJSON();
 
-            //         let response1 = await axios({
-            //             method: requestData.config.method,
-            //             url: requestData.config.baseURL + requestData.config.url,
-            //             headers: { "content-type": "application/json", "authorization": getAccessToken() },
-            //             params: requestData.config.params
-            //         });
-            //     }
-            // } catch (error) {
-            //     return Promise.reject(error)
-            // }
+                    let response1 = await axios({
+                        method: requestData.config.method,
+                        url: requestData.config.baseURL + requestData.config.url,
+                        headers: { "content-type": "application/json", "authorization": getAccessToken() },
+                        params: requestData.config.params
+                    });
+                }
+            } catch (error) {
+                return Promise.reject(error)
+            }
         } else {
             console.log("ERROR IN RESPONSE: ", error.toJSON());
             return {
